Validate victory Quest and state tracing

+++ src/gameplay/rng.c
@@ -4,12 +4,9 @@
  *   FUN_8001714c (V8_SeedRng)  -- writes g_rngSeed and zeroes the counter
  *                                  at @ 0x800568d4 / 0x800568d8.
  *
- * Pass 2 must locate the actual rand() consumer (LW from 0x800568d4 with
- * a multiply-add update). For now we have the seeder only.
- *
  * Bit-exact-critical: see PROJECT_SCOPE.md "Random Number Generation".
  * AI behavior, weapon spread, and quest events depend on the exact
- * algorithm. We MUST decompile the consumer side in pass 2.
+ * algorithm. Both the seeder and consumer below are instruction-confirmed.
  *
  * Known caller: FUN_80015098 (main) sets seed = 0xbb40e64d when
  * g_matchMode == MATCH_MODE_SPECIAL5. This is a *deterministic* seed --
@@ -19,7 +16,7 @@
 
 /* HIGH: globals at fixed VA. */
 int32_t g_rngSeed;     /* @ 0x800568d4 */
-int32_t g_rngCounter;  /* @ 0x800568d8 */
+uint8_t g_rngCounter;  /* @ 0x800568d8: LBU/SB one-byte carry */
 
 /* HIGH: seed the RNG. Always pairs (seed, counter=0). */
 void V8_SeedRng(int32_t seed)
@@ -53,8 +50,8 @@ void V8_SeedRng(int32_t seed)
  */
 uint32_t V8_RandNext(void)
 {
-    uint32_t prev = (uint32_t)(uint8_t)g_rngCounter;
-    g_rngCounter = (int32_t)(uint8_t)g_rngSeed;
+    uint32_t prev = g_rngCounter;
+    g_rngCounter = (uint8_t)g_rngSeed;
     uint32_t v   = ((uint32_t)g_rngSeed >> 1) | (prev << 31);
     v ^= ((uint32_t)g_rngSeed << 12);
     g_rngSeed = (int32_t)(v ^ (v >> 20));
